requirements.txt
# Core Dependencies
fastapi==0.110.0
uvicorn==0.23.2
pydantic==2.11.3
pydantic-core==2.33.1
pydantic-settings==2.8.1
python-dotenv==1.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.3.0
cryptography==44.0.2
email_validator==2.2.0
python-multipart==0.0.20
anyio>=3.7.1,<5
starlette==0.36.3
h11==0.14.0
gunicorn==23.0.0

# Database
pymongo==4.12.0
dnspython==2.7.0

# WebSocket
websockets==14.2
httptools==0.6.4
watchfiles==1.0.5
#uvloop==0.21.0
wsproto==1.2.0

# Base ML Dependencies
numpy==1.26.4
scipy==1.15.2
scikit-learn==1.6.1
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
pandas==2.2.3
matplotlib==3.10.1

# Transformers & NLP
tokenizers==0.21.1
huggingface-hub==0.30.2
transformers==4.51.1
sentence-transformers==4.0.2
keybert==0.7.0
nltk==3.9.1
regex==2024.11.6
sentencepiece==0.2.0

# Audio Processing
pydub==0.25.1
sounddevice==0.5.1
librosa==0.11.0
audioread==3.0.1
soundfile==0.13.1
numba==0.61.2
pooch==1.8.2
lazy_loader==0.4
soxr==0.5.0.post1
llvmlite==0.44.0

# Computer Vision
opencv-python==4.7.0.72
dlib==19.24.6
imutils==0.5.4
deepface==0.0.93
retina-face==0.0.17
mtcnn==1.0.0
pillow==11.1.0

# Google AI
google-generativeai==0.3.1
google-genai==1.10.0
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.2
protobuf==4.25.6
google-ai-generativelanguage==0.4.0

# TensorFlow
tensorflow==2.19.0
tensorflow-datasets==4.9.3
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.15.0
absl-py==2.2.2
astunparse==1.6.3
flatbuffers==25.2.10
gast==0.6.0
grpcio==1.72.0rc1
grpcio-status==1.63.0rc1
h5py==3.13.0
keras==3.9.2
opt_einsum==3.4.0
termcolor==3.0.1
wrapt==1.17.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tf_keras==2.19.0

# HTTP & API
httpcore==1.0.7
httpx==0.28.1
requests==2.32.3
urllib3==2.3.0

# Web Development
Flask==3.1.0
flask-cors==5.0.1
Werkzeug==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2

# Utilities
tqdm==4.67.1
typing_extensions==4.13.1
typing-inspection==0.4.0
six==1.17.0
packaging==24.2
filelock==3.18.0
joblib==1.4.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
rsa==4.9
setuptools==78.1.0
wheel==0.44.0
rich==13.8.1
sniffio==1.3.1
idna==3.10
annotated-types==0.7.0
ecdsa==0.19.1
pycparser==2.22
safetensors==0.5.3
sympy==1.13.1
mpmath==1.3.0
networkx==3.4.2
PyYAML==6.0.2
beautifulsoup4==4.13.3
soupsieve==2.6
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.3.2
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
fonttools==4.57.0
fsspec==2024.9.0
kiwisolver==1.4.8
markdown-it-py==3.0.0
mdurl==0.1.2
platformdirs==4.3.7
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
threadpoolctl==3.6.0
tzdata==2025.2

python-pptx==0.6.21
google-generativeai==0.3.1